First attempt at automating blast search
not very efficient - uses procedural programming and calls blast
using back ticks to do command line things. Works, but gives
lots of files that have to be sorted through and thisis unwieldy
if the searching with lots of sequences.

In principal making lots of blast read outs and then searching
through using regular expressions to find the hits isn't unreasonable.