Blast against Eurydice genome using BioPerl
This is a better way to do the blasting.

Using the bioperl wrapper for blast and seqIO objects from bioperl this is
a much quicker script.

This does the blast and telling whether there is a hit or not in one script
but some information is lost - I didn't keep the blast reports so I
couldn't say how good of a match it is.