fixed bug that prevented proper reports
using Align::IO to give pretty alignments in the individual reports
caused only the first HSP to be printed because the writing to file
function I used closed the file handle after the first one was
printed.

To fix this I stopped using the pretty alignment and printed the
ugly, basic one from SearchIO. Looks much uglier as each string is
on its own line and if the match is long then it soft wraps or goes
miles off the edge of the screen. However all the HSPs are printed,
so over all a win IMO.